+++ packages/go-kosu/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+-   Fix Order store I/O overhead
 -   Fix wrong initial validators power calculations
 -   Add --validator family flags to optionally start as a validator node
 -   Add GetBlocks RPC endpoint